Fix compiler warning C4245: signed/unsigned mismatch when converting between types.  Patch by Sorin Jianu.  r=me,bryner
git-svn-id: http://google-breakpad.googlecode.com/svn/trunk@217 4c0a9323-5329-0410-9bdc-e9ce6186880e

+++ src/common/windows/http_upload.cc
@@ -137,7 +137,8 @@ bool HTTPUpload::SendRequest(const wstring &url,
   wstring content_type_header = GenerateRequestHeader(boundary);
   HttpAddRequestHeaders(request.get(),
                         content_type_header.c_str(),
-                        -1, HTTP_ADDREQ_FLAG_ADD);
+                        static_cast<DWORD>(-1),
+                        HTTP_ADDREQ_FLAG_ADD);
 
   string request_body;
   if (!GenerateRequestBody(parameters, upload_file,
